chore: move repository and service module to HandPoint package.
