+++ app/src/main/java/com/ishzk/android/majancalculator/ui/HandPointViewModel.kt
@@ -1,6 +1,10 @@
 package com.ishzk.android.majancalculator.ui
 
 import androidx.lifecycle.ViewModel
+import com.ishzk.android.majancalculator.domain.handpoint.PointRepository
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
 
-class HandPointViewModel: ViewModel() {
+@HiltViewModel
+class HandPointViewModel @Inject constructor(private val repository: PointRepository): ViewModel() {
 }
